Move IQueryStep to datasource module
