+++ server/datasource/common/src/main/java/org/bithon/server/datasource/query/pipeline/IQueryStep.java
@@ -1,0 +1,31 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.server.datasource.query.pipeline;
+
+
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * @author frank.chen021@outlook.com
+ * @date 4/4/25 3:47 pm
+ */
+public interface IQueryStep {
+
+    boolean isScalar();
+
+    CompletableFuture<PipelineQueryResult> execute() throws Exception;
+}
