+++ server/metric-expression/pom.xml
@@ -19,6 +19,10 @@
 
         <dependency>
             <groupId>org.bithon.server</groupId>
+            <artifactId>datasource-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.bithon.server</groupId>
             <artifactId>server-web-service</artifactId>
         </dependency>
         <dependency>
