+++ server/metric-expression/src/main/java/org/bithon/server/metric/expression/api/MetricQueryApi.java
@@ -24,12 +24,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.bithon.component.commons.Experimental;
-import org.bithon.server.metric.expression.pipeline.IQueryStep;
+import org.bithon.server.datasource.query.pipeline.Column;
+import org.bithon.server.datasource.query.pipeline.IQueryStep;
+import org.bithon.server.datasource.query.pipeline.PipelineQueryResult;
 import org.bithon.server.metric.expression.pipeline.QueryPipelineBuilder;
 import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.bithon.server.web.service.datasource.api.IDataSourceApi;
 import org.bithon.server.web.service.datasource.api.IntervalRequest;
 import org.bithon.server.web.service.datasource.api.QueryResponse;
+import org.bithon.server.web.service.datasource.api.TimeSeriesMetric;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -38,6 +41,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Nullable;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author frank.chen021@outlook.com
@@ -87,5 +94,57 @@ public class MetricQueryApi {
         return evaluator.execute()
                         .get()
                         .toTimeSeriesResultSet();
+    }
+
+    private QueryResponse<?> toTimeSeriesResultSet(PipelineQueryResult intermediateResult) {
+        // Because the end timestamp is inclusive, we need to add 1
+        int count = 1 + (int) ((intermediateResult.getEndTimestamp() - intermediateResult.getStartTimestamp()) / intermediateResult.getInterval());
+
+        List<String> keys = new ArrayList<>(intermediateResult.getKeyColumns());
+        if (keys.get(0).equals("_timestamp")) {
+            keys.remove(0);
+        } else {
+            throw new IllegalStateException();
+        }
+        Column timestampCol = intermediateResult.getTable().getColumn("_timestamp");
+        List<Column> dimCols = intermediateResult.getTable().getColumns(keys);
+        List<Column> valCols = intermediateResult.getTable().getColumns(intermediateResult.getValColumns());
+
+        Map<List<String>, TimeSeriesMetric> map = new LinkedHashMap<>(7);
+        for (int i = 0; i < intermediateResult.getRows(); i++) {
+
+            // the timestamp is seconds
+            long timestamp = timestampCol.getLong(i) * 1000;
+            long index = (timestamp - intermediateResult.getStartTimestamp()) / intermediateResult.getInterval();
+
+            for (int j = 0, valColsSize = valCols.size(); j < valColsSize; j++) {
+                Column valCol = valCols.get(j);
+                List<String> series = new ArrayList<>(dimCols.size() + 1);
+                for (Column dimCol : dimCols) {
+                    series.add(dimCol.getObject(i).toString());
+                }
+                series.add(intermediateResult.getValColumns().get(j));
+
+                map.computeIfAbsent(series, k -> new TimeSeriesMetric(series, count))
+                   .set((int) index, valCol.getObject(i));
+            }
+        }
+
+
+        List<QueryResponse.QueryResponseColumn> responseColumns = new ArrayList<>(dimCols.size() + valColumns.size());
+        for (Column dim : dimCols) {
+            responseColumns.add(new QueryResponse.QueryResponseColumn(dim.getName(), dim.getDataType().name()));
+        }
+        for (Column val : valCols) {
+            responseColumns.add(new QueryResponse.QueryResponseColumn(val.getName(), val.getDataType().name()));
+        }
+
+        return QueryResponse.builder()
+                            .interval(intermediateResult.getInterval())
+                            .startTimestamp(intermediateResult.getStartTimestamp())
+                            .endTimestamp(intermediateResult.getEndTimestamp())
+                            .data(map.values())
+                            .meta(responseColumns)
+                            .build();
     }
 }
