+++ server/metric-expression/src/main/java/org/bithon/server/metric/expression/pipeline/QueryPipelineBuilder.java
@@ -21,10 +21,14 @@ import org.bithon.component.commons.expression.ArithmeticExpression;
 import org.bithon.component.commons.expression.IExpression;
 import org.bithon.component.commons.expression.LiteralExpression;
 import org.bithon.component.commons.utils.StringUtils;
+import org.bithon.server.datasource.query.pipeline.IQueryStep;
 import org.bithon.server.metric.expression.ast.IMetricExpressionVisitor;
 import org.bithon.server.metric.expression.ast.MetricExpression;
 import org.bithon.server.metric.expression.ast.MetricExpressionASTBuilder;
 import org.bithon.server.metric.expression.ast.MetricExpressionOptimizer;
+import org.bithon.server.metric.expression.pipeline.step.BinaryExpressionQueryStep;
+import org.bithon.server.metric.expression.pipeline.step.LiteralQueryStep;
+import org.bithon.server.metric.expression.pipeline.step.MetricExpressionQueryStep;
 import org.bithon.server.web.service.datasource.api.IDataSourceApi;
 import org.bithon.server.web.service.datasource.api.IntervalRequest;
 import org.bithon.server.web.service.datasource.api.QueryField;
