+++ server/metric-expression/src/main/java/org/bithon/server/metric/expression/pipeline/step/HashJoiner.java
@@ -14,9 +14,11 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.metric.expression.format;
+package org.bithon.server.metric.expression.pipeline.step;
 
 import org.bithon.component.commons.utils.CollectionUtils;
+import org.bithon.server.datasource.query.pipeline.Column;
+import org.bithon.server.datasource.query.pipeline.ColumnarTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
