+++ server/metric-expression/src/main/java/org/bithon/server/metric/expression/pipeline/step/LiteralQueryStep.java
@@ -14,13 +14,15 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.metric.expression.pipeline;
+package org.bithon.server.metric.expression.pipeline.step;
 
 
 import org.bithon.component.commons.expression.IDataType;
 import org.bithon.component.commons.expression.LiteralExpression;
-import org.bithon.server.metric.expression.format.Column;
-import org.bithon.server.metric.expression.format.ColumnarTable;
+import org.bithon.server.datasource.query.pipeline.Column;
+import org.bithon.server.datasource.query.pipeline.ColumnarTable;
+import org.bithon.server.datasource.query.pipeline.IQueryStep;
+import org.bithon.server.datasource.query.pipeline.PipelineQueryResult;
 
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -44,7 +46,7 @@ public class LiteralQueryStep implements IQueryStep {
     }
 
     @Override
-    public CompletableFuture<IntermediateQueryResult> execute() {
+    public CompletableFuture<PipelineQueryResult> execute() {
 
         ColumnarTable table = new ColumnarTable();
         if (expression.getDataType() == IDataType.LONG) {
@@ -61,10 +63,10 @@ public class LiteralQueryStep implements IQueryStep {
             throw new IllegalStateException("Unsupported literal type: " + expression.getDataType());
         }
 
-        return CompletableFuture.completedFuture(IntermediateQueryResult.builder()
-                                                                        .rows(1)
-                                                                        .valColumns(List.of("value"))
-                                                                        .table(table)
-                                                                        .build());
+        return CompletableFuture.completedFuture(PipelineQueryResult.builder()
+                                                                    .rows(1)
+                                                                    .valColumns(List.of("value"))
+                                                                    .table(table)
+                                                                    .build());
     }
 }
